* Version: 1.2.16 	* GormInfo.plist 	* Documentation/news.texi: Update version and documentation.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@34723 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassInspector.m
@@ -312,7 +312,7 @@ objectValueForTableColumn: (NSTableColumn *)tc
 
   NSArray *list = [classManager allClassNames];
   NSString *superClass = [classManager parentOfClass: [self _currentClass]];
-  int index = [list indexOfObject: superClass];
+  NSUInteger index = [list indexOfObject: superClass];
 
   [classField setStringValue: [self _currentClass]];
   [outletTable reloadData];
